% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_adi.R
\name{calc_adi_dry}
\alias{calc_adi_dry}
\title{Calculate the final temperature by dry adiabatic lapse rate}
\usage{
calc_adi_dry(data, pres1, temp1, pres2, name_as = "temp2", overwrite = F)
}
\arguments{
\item{overwrite}{}
}
\value{

}
\description{
Calculate the final temperature by dry adiabatic lapse rate
}
