% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_adi.R
\name{calc_adi_wet}
\alias{calc_adi_wet}
\title{Calculate the final temperature/ pressure by wet adiabatic lapse rate}
\usage{
calc_adi_wet(
  data,
  pres1,
  temp1,
  pres2,
  acc = 1,
  name_as = c("pres2", "temp2"),
  overwrite = F
)
}
\arguments{
\item{overwrite}{}
}
\value{

}
\description{
Calculate the final temperature/ pressure by wet adiabatic lapse rate
}
