% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ccl_est.R
\name{calc_ccl_est}
\alias{calc_ccl_est}
\title{Calculate the \strong{estimated} final temperature and pressure of Convective Condensation Level}
\usage{
calc_ccl_est(
  data,
  pres,
  temp,
  dwpt,
  ...,
  name_as = c("cclp", "cclt", "cclh"),
  overwrite = F
)
}
\arguments{
\item{data}{The dataframe itself. Each row of data will be considered as an observation for each calculation}

\item{pres}{The column name of the initial pressure of the calculation. In any units.}

\item{temp}{The column name of the initial temperature of the calculation. Unit in \code{degC}.}

\item{dwpt}{The column name of the initial dewpoint of the calculation. Unit in \code{degC}.}

\item{...}{The column name of grouping of data}

\item{name_as}{Names of the 3 new columns, in the order of "pressure of the CCL", "temperature of the CCL(estimated by smoothened temperature profile)", and "temperature of the CCL(estimated by isohume)". Default as \code{c("cclp", "cclt", "cclh")}.}

\item{overwrite}{Let the new column names to overwrite the original dataframe columns? Default as \code{FALSE}.}
}
\description{
The Convective Condensation Level, or \strong{CCL}, refers to "the height (or pressure) where an air parcel becomes saturated when heated from below and lifted adiabatically due to buoyancy.". Or, in other words, CCL is when the isohume of an air parcel intersects with temperature profile of a sounding.
Typically, CCL is only used for air parcels from the surface, as the pressure level of CCL must be less than the starting surface in order to cool adiabatically. However, for this calculator, CCL is calculated as the closest pressure level that the corresponding isohume intersects with the temperature profile, allowing for multi-level calculation of CCL.
Do note that this algorithum supports computate of multiple sounding at once. \strong{However, multiple sounding must be grouped by additional columns.} Feed the additional columns used in the \code{...} argument.
}
\examples{
load_asnd(time = Sys.time()) \%>\% calc_ccl_est(pres, temp, dwpt, station, time)
}
