% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_td_ema.R
\name{calc_emay}
\alias{calc_emay}
\title{Calculate Emagram plot y-coordinate}
\usage{
calc_emay(data, y, name_as = NULL)
}
\arguments{
\item{name_as}{}
}
\description{
Calculate Emagram plot y-coordinate
}
\keyword{internal}
