% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_geo.R
\name{calc_geo_dest}
\alias{calc_geo_dest}
\title{Calculate location by distance and bearing}
\usage{
calc_geo_dest(
  data,
  lon1,
  lat1,
  brn1,
  dis,
  name_as = c("lon2", "lat2", "brn2", "error"),
  overwrite = F,
  write_error = F,
  acc = 1e-12,
  a = 6378137,
  b = 6356752.314245
)
}
\arguments{
\item{data}{The data}

\item{lon1}{The column name of the longitude (in deg)}

\item{lat1}{The column name of the latitude (in deg)}

\item{brn1}{The column name of the bearning (in deg)}

\item{dis}{The column name of the distance (in m)}

\item{name_as}{The names of the new column name. Default as \code{c("lon2", "lat2", "brn2", "error")}.}

\item{overwrite}{Let the new column names to overwrite the original dataframe columns? Default as \code{FALSE}.}

\item{write_error}{Do you write the \code{error} column showing the estimation error? Default as \code{FALSE}.}

\item{acc}{Minimum accuracy of the estimation. Default as \code{1e-12}}

\item{a}{Major semi-axes of the ellipsoid, i.e. mean diameter of the Earth's Equator. \strong{DO NOT CHANGE IF YOU DON'T KNOW WHAT YOU ARE DOING}}

\item{b}{Minor semi-axes of the ellipsoid, i.e. mean diameter of the Earth's Prime Meridian. \strong{DO NOT CHANGE IF YOU DON'T KNOW WHAT YOU ARE DOING}}
}
\description{
Using the Vincenty direct solution of geodesics on the ellipsoid.
A R adaptation using https://www.movable-type.co.uk/scripts/latlong-vincenty.html code.
}
\examples{
calc_geo_dest(data = data.frame(lon1 = 1:200, lat1 = 1:200, brn1 = 1:200, dis = seq(1100, 21000, 100)), lon1, lat1, brn1, dis, write_error = T)
}
