% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_adi.R
\name{calc_isohume}
\alias{calc_isohume}
\title{Calculate the final temperature by saturated mixing ratio}
\usage{
calc_isohume(data, pres1, temp1, pres2, name_as = "temp2", overwrite = F)
}
\arguments{
\item{overwrite}{}
}
\value{

}
\description{
Calculate the final temperature by saturated mixing ratio
}
