% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lcl.R
\name{calc_lcl}
\alias{calc_lcl}
\title{Calculate the final temperature and pressure of Lifting Condensation Level, with high accurary}
\usage{
calc_lcl(
  data,
  pres,
  temp,
  dwpt,
  acc = 5,
  name_as = c("lclp", "lclt_t", "lclt_d"),
  overwrite = F
)
}
\arguments{
\item{data}{The dataframe itself. Each row of data will be considered as an observation for each calculation}

\item{pres}{The column name of the initial pressure of the calculation. In any units.}

\item{temp}{The column name of the initial temperature of the calculation. Unit in \code{degC}.}

\item{dwpt}{The column name of the initial dewpoint of the calculation. Unit in \code{degC}.}

\item{name_as}{Names of the 3 new columns, in the order of "pressure of the LCL", "temperature of the LCL(estimated by temperature)", and "temperature of the LCL(estimated by dewpoint)". Default as \code{c("lclp", "lclt_t", "lclt_d")}. Keyword \code{"*del*"} is supported.}

\item{overwrite}{Let the new column names to overwrite the original dataframe columns? Default as \code{FALSE}.}
}
\value{

}
\description{
Calculate the final temperature and pressure of Lifting Condensation Level, with high accurary
}
