% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lcl.R
\name{calc_lcl_est}
\alias{calc_lcl_est}
\title{Calculate the \strong{estimated} final temperature and pressure of Lifting Condensation Level}
\usage{
calc_lcl_est(data, pres, temp, dwpt, name_as = NULL, overwrite = F)
}
\arguments{
\item{data}{The dataframe itself. Each row of data will be considered as an observation for each calculation.}

\item{pres}{The column name of the air parcel pressure level. Unit in \code{hPa}.}

\item{temp}{The column name of the air parcel temperature. Unit in \code{degC}.}

\item{dwpt}{The column name of the air parcel dewpoint. Unit in \code{degC}.}

\item{name_as}{Names of the 3 new columns, in the order of "pressure of the LCL", "temperature of the LCL(estimated by temperature)", and "temperature of the LCL(estimated by dewpoint)". Default as \code{NULL}, i.e. the column names of \code{pres}, \code{temp}, and \code{dwpt}, with a prefix of \code{lcl}. Keyword \code{"*del*"} is supported.}

\item{overwrite}{Let the new column names to overwrite the original dataframe columns? Default as \code{FALSE}.}
}
\value{
The same dataframe as \code{data}, with 3 additional column with column name defined in \code{name_as}.
}
\description{
Lifting Condensation Level (lcl) refers to the pressure level where relative humidity of an air parcel will reach 100\% with respect to liquid water when it's cooled by dry adiabatic lifting, i.e. only due to buoyancy of the air parcel.
}
\details{
Understanding lcl is crucial in meteorology, as this shows where clouds/ fog may start to form. This formula calculates the lcl of an air parcel, given its pressure (\code{pres}), temperature (\code{temp}) and dew point (\code{dwpt}) are known
}
\examples{
calc_lcl_est(tibble::tibble(pres = 1000, temp = 20, dwpt = 10), pres, temp, dwpt) #Should return 861 for pressure
}
