% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_parcel.R
\name{calc_parcel}
\alias{calc_parcel}
\title{Calculate the final temperature of an air parcel.}
\usage{
calc_parcel(
  data,
  pres1,
  temp1,
  dwpt1,
  pres2,
  acc = 1,
  name_as = NULL,
  overwrite = F
)
}
\arguments{
\item{data}{The dataframe itself. Each row of data will be considered as an observation for each calculation.}

\item{pres1}{The column name of the initial pressure of the calculation. Unit in \code{hPa}.}

\item{temp1}{The column name of the initial temperature of the calculation. Unit in \code{degC}.}

\item{dwpt1}{The column name of the initial dewpoint of the calculation. Unit in \code{degC}.}

\item{pres2}{The column name of the final pressure of the calculation. Unit in \code{hPa}.}

\item{acc}{Accuracy of the calculation if uses moist adiabetic rate, which will be fed into \code{calc_adi_wet}.}

\item{name_as}{Names of the 2 new columns, in the order of "temperature of the final calculation", and "dewpoint of the final calculation". Default as \code{NULL}, which will utilized the column name of \code{temp1} and \code{dwpt1}, with suffix \code{"2"}. Keyword \code{"*del*"} is supported.}

\item{overwrite}{Let the new column names to overwrite the original dataframe columns? Default as \code{FALSE}.}
}
\description{
An air parcel can be dry-adiabetically and moist-adiabetically cooled/ heated, which is determined by the mixing ratio of the said parcel. This formula automatically selects the correct final temperature and dew point of the parcel, by \code{calc_lcl} first.
}
\examples{
weather2::load_uw_asnd() \%>\%
dplyr::slice(1) \%>\%
dplyr::select(pres, temp, dwpt) \%>\%
dplyr::mutate(pres2 = 200) \%>\%
weather2::calc_parcel(pres, temp, dwpt, pres2)
}
