% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_smooth.R
\name{calc_smooth}
\alias{calc_smooth}
\title{Calculate smoothing and filling data}
\usage{
calc_smooth(
  data,
  based,
  value,
  name_as,
  type,
  weight = rep(1, 5),
  fill = seq(1, 50000, 1),
  df = 1
)
}
\arguments{
\item{data}{Atmospheric sounding data}

\item{based}{Parameter used for smoothing, often based on height, but can also be based on pressure or other parameters. Must ensure there is no NA values.}

\item{value}{Value to be smoothed}

\item{name_as}{Column name of the smoothed data.}

\item{type}{Type of smoothing. Only accepts "moving", "linear", and "spline".}

\item{weight}{Weight for smoothing. Only used when type is "moving".}

\item{fill}{Data to be filled based on 'based'.}

\item{df}{Degree of freedom. Only used when type is "spline".}
}
\value{

}
\description{
Calculate smoothing and filling data
}
\examples{
calc(data, hght, uwnd, type = "moving")
}
