% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_smooth_ma.R
\name{calc_smooth_ma}
\alias{calc_smooth_ma}
\title{Calculate by smoothing the column, by moving average}
\usage{
calc_smooth_ma(data, based, value, side = "center", weight = 3)
}
\arguments{
\item{data}{The Data frame}

\item{based}{A column within dataframe. This column must be "numeric-able", must not contain NAs, and must be unique}

\item{value}{A column within data}

\item{side}{Accepts 1 string of "left", "center", or "right"}

\item{weight}{weighting or the size of the window}
}
\value{

}
\description{
Calculate by smoothing the column, by moving average
}
\examples{
calc_smooth_ma(data, x, y)
}
