% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/const.R
\name{const}
\alias{const}
\title{Calling a scientific constant}
\usage{
const(name, value, unit, silent = F)
}
\arguments{
\item{name}{Name of the constant}

\item{value}{Value of the constant}

\item{unit}{Unit of the constant}

\item{silent}{Should a message be returned when calling the constant? Default as \code{FALSE}.}
}
\description{
Calling a scientific constant
}
\examples{
const("standard acceleration of gravity", 9.80665, "m s^-2")
}
\keyword{internal}
