% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_radar.R
\name{coord_radar}
\alias{coord_radar}
\title{ggplot2: Radar Coordiante System}
\usage{
coord_radar(theta = "x", start = 0, direction = 1, ...)
}
\arguments{
\item{theta}{Variable to map angle to x or y}

\item{start}{Offset of starting point from 12 o'clock in radians. Offset is applied clockwise or anticlockwise depending on value of direction.}

\item{direction}{1, clockwise; -1, anticlockwise}

\item{...}{Other arguments to be passed to ggproto}
}
\value{

}
\description{
Adds a radar coordinate system for radar charts. Inspired by see
}
