% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_asnd.R
\name{draw_asnd}
\alias{draw_asnd}
\title{Plot atmospheric sounding data}
\usage{
draw_asnd(
  data,
  pres,
  temp,
  dwpt,
  frpt,
  drct,
  sped,
  hght,
  thta,
  thte,
  thtv,
  log = T,
  ybreak = seq(100, 1000, 100),
  minor_ybreak = seq(100, 1000, 50),
  xbreak = seq(-100, +60, 10),
  minor_xbreak = seq(-100, 60, 5),
  sspan = 0.2,
  sbreak = seq(0, 360, 90)
)
}
\arguments{
\item{data}{Dataframe or Weather object of atmospheric sounding data}

\item{pres}{column name of atmospheric pressure, must be provided}

\item{temp}{column name of air temperature, must be provided}

\item{dwpt}{column name of dew point}

\item{frpt}{column name of height}

\item{drct}{column name of wind direction}

\item{sped}{column name of wind speed}

\item{hght}{column name for geopotential height}

\item{thta}{column name of potential temperature}

\item{thte}{column name of equivalent potential temperature}

\item{thtv}{column name of virtual potential temperature}

\item{log}{log status of the y-axis. Only accept logic values (T/F). T for log-10 scale. F for linear scale.}

\item{ybreak}{Breaks for the y-axis.}

\item{minor_ybreak}{Minor breaks for the y-axis.}

\item{xbreak}{Breaks for the x-axis.}

\item{minor_xbreak}{Minor breaks for the x-axis.}

\item{sspan}{Size of the secondary x-axis.}

\item{sbreak}{Breaks for the secondary x-axis.}
}
\value{

}
\description{
Plot atmospheric sounding data
}
