% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_hodo.R
\name{draw_hodo}
\alias{draw_hodo}
\title{Plot sounding hodograph}
\usage{
draw_hodo(
  data,
  uwnd,
  vwnd,
  pres,
  hght,
  color = "hght",
  label_breaks,
  label_text = F,
  seed = NA,
  breaks = seq(-100, 100, 10),
  center = "both",
  equal = T
)
}
\arguments{
\item{data}{Dataframe or Weather object of atmospheric sounding data}

\item{uwnd}{column name of u-wind vector}

\item{vwnd}{column name of v-wind vector}

\item{pres}{column name of pressure}

\item{hght}{column name of height}

\item{color}{parameter for the hodograph line color. "hght" for hodograph line to be based on height. "pres" for hodograph line to be based on pressure.}

\item{label_breaks}{position of label breaks. must provide "hght" or "pres"}

\item{label_text}{switch for texted label. Only accepts T/F values}

\item{seed}{Seed for the label text to avoid overlapping}

\item{breaks}{breaks for the x and y-axis}

\item{center}{parameter for centering the graph at 0. Accpets "both", "uwnd", "vwnd", and both "uwnd" and "vwnd" at the same time.}

\item{equal}{switch for equal scaling between x-axis and y-axis.}
}
\value{
a ggplot based hodoplot
}
\description{
Plot sounding hodograph
}
\examples{
draw_hodo(data, uwnd, vwnd, pres, hght, color = "pres", seed = 1, label_breaks = c("hght", seq(0, 20000, 10)), label_text = T)
}
