% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_rose.R
\name{draw_rose}
\alias{draw_rose}
\title{Plot rose graph}
\usage{
draw_rose(data, drct, value, sum, breaks = 45, minor_breaks)
}
\arguments{
\item{data}{data frame containing data}

\item{drct}{column name of direction}

\item{value}{column name of value}

\item{sum}{column name of sum (optional)}

\item{breaks}{Interval between each break. Default as 45.}

\item{minor_breaks}{Interval between each minor break. Default as half of "breaks".}
}
\value{
A \code{ggplot2} based roseplot
}
\description{
\strong{NOTE: This is to-be-depreciated soon! Use the to-be new function \code{plot_rose} instead}
}
\examples{
draw_rose(data, drct, sped, time_int, breaks = 30, minor_breaks = 10)
}
