% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_list.R
\name{file_list}
\alias{file_list}
\title{List a list of files with a certain pattern, within a directory (and it's subdirectory)}
\usage{
file_list(dir, pattern, subdirectory = T)
}
\arguments{
\item{dir}{a list of directories}

\item{pattern}{a list of patterns to look for. Uses regular expression}

\item{subdirectory}{Include subdirectory or not}
}
\description{
List a list of files with a certain pattern, within a directory (and it's subdirectory)
}
\examples{
file_list("C:/Users/carol/Desktop", "*_hi.csv")
}
