% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_list.R
\name{file_list_csv}
\alias{file_list_csv}
\title{List a list of .csv files within a directory (and it's subdirectory)}
\usage{
file_list_csv(dir, subdirectory = T)
}
\arguments{
\item{dir}{a list of directories}

\item{subdirectory}{Include subdirectory or not}
}
\value{

}
\description{
List a list of .csv files within a directory (and it's subdirectory)
}
\examples{
file_list_csv("C:/Users/carol/Desktop")
}
