% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_list.R
\name{file_list_gif}
\alias{file_list_gif}
\title{List a list of *.gif files within a directory (and it's subdirectory)}
\usage{
file_list_gif(dir, subdirectory = T)
}
\arguments{
\item{dir}{a list of directories}

\item{subdirectory}{Include subdirectory or not}
}
\description{
List a list of *.gif files within a directory (and it's subdirectory)
}
\examples{
file_list_gif("C:/Users/carol/Desktop")
}
