% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_sys_info_windows.R
\name{file_sysinfo_windows}
\alias{file_sysinfo_windows}
\title{Get the Size and FreeSpace of Windows machine}
\usage{
file_sysinfo_windows()
}
\description{
Get the Size and FreeSpace of Windows machine
}
\examples{
file_sysinfo_windows()
}
