% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_load_wxph.R
\name{hk_load_wxph}
\alias{hk_load_wxph}
\title{HK weather - Downloading weather photo from https://www.hko.gov.hk/en/index.html}
\usage{
hk_load_wxph(
  time = weather2::tool_datetime(end = Sys.time(), by = "10 min", duration = "3 day"),
  station = "all",
  list_fail = F,
  dir = getwd(),
  attempt = 10,
  worker = 1
)
}
\arguments{
\item{time}{Date/time to download}

\item{list_fail}{List failed-to-download items}

\item{dir}{Directory of downloaded data.}

\item{attempt}{Attempts to be made per file to download.}

\item{worker}{Numbers of sessions to be open}

\item{magn}{Satellite image magnification. Accepts 2/ 4/ 8, or any combinations}

\item{type}{Type of satellite image. Accepts "tc" (True color), "ir" (Infrared), "dc" (Deep convection), or any combinations.}
}
\value{

}
\description{
HK weather - Downloading weather photo from https://www.hko.gov.hk/en/index.html
}
\examples{
hk_load_wxph()
}
