% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_file_gif.R
\name{list_file_gif}
\alias{list_file_gif}
\title{List a list of *.gif files within a directory (and it's subdirectory)}
\usage{
list_file_gif(dir, subdirectory = T)
}
\arguments{
\item{dir}{a list of directories}

\item{subdirectory}{Include subdirectory or not}
}
\value{

}
\description{
List a list of *.gif files within a directory (and it's subdirectory)
}
\examples{
list_file_gif("C:/Users/carol/Desktop")
}
