% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_asnd.R
\name{load_asnd}
\alias{load_asnd}
\title{Download atmospheric sounding data}
\usage{
load_asnd(
  time = tool_datetime(end = Sys.time(), duration = "1 days", by = "12 hours"),
  station = 45004,
  frost = T,
  attempt = 50
)
}
\arguments{
\item{time}{A list of time. If said time does not have a sounding, it will try to adjust to the last sounding possible. Only accepts POSIXct.}

\item{station}{Weather station code, according to the University of Wyoming website.}

\item{frost}{Include frost point calculation in dataframe. Only accepts logical value.}

\item{attempt}{Attempts of downloads to try per time. Default as 50.}
}
\value{

}
\description{
Understanding the sounding is important for understanding the atmospheric condition, especially for extreme weather.
This catches the sounding data provided by the University of Wyoming.
}
\examples{
load_asnd()
}
