% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_asnd_metr.R
\name{load_asnd_metr}
\alias{load_asnd_metr}
\title{Download atmospheric sounding data metrics}
\usage{
load_asnd_metr(
  time = tool_datetime(end = Sys.time(), duration = "7 days", by = "12 hours"),
  station = 45004,
  attempt = 50
)
}
\arguments{
\item{time}{A list of time. If said time does not have a sounding, it will try to adjust to the last sounding possible. Only accepts POSIXct.}

\item{station}{Weather station code, according to the University of Wyoming website.}

\item{attempt}{Attempts of downloads to try per time. Default as 50.}
}
\description{
Understanding the sounding is important for understanding the atmospheric condition, especially for extreme weather.
This catches the metrics provided by the University of Wyoming.
}
\examples{
load_asnd_metr()
}
