% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_metar.R
\name{load_metar}
\alias{load_metar}
\title{Download METAR data}
\usage{
load_metar(
  time = tool_date(end = as.Date(lubridate::with_tz(Sys.time(), tzone = "UTC")), by =
    "1 days", duration = "1 days"),
  station = "VHHH",
  attempt = 50,
  decode = T,
  CAVOK = T
)
}
\arguments{
\item{time}{Date to be download (UTC time)}

\item{station}{ICAO code}

\item{attempt}{Attempts of downloads to try per time. Default as 50.}

\item{decode}{Decode the METAR code automatically. Default as T.}

\item{CAVOK}{If decode is T, supress CAVOK and convert all CAVOK into vsby = 9999. Default as T}
}
\value{

}
\description{
Meteorological Aviation Report, or METAR, are weather reports issued at airports for aviators.
These reports have a high temporal resolution, updating at least once per hour.
}
\examples{
load_metar()
}
