% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_uw_asnd.R
\name{load_uw_asnd}
\alias{load_uw_asnd}
\title{Download atmospheric sounding data}
\usage{
load_uw_asnd(
  time = Sys.time(),
  station = "45004",
  frost = T,
  attempt = 5L,
  worker = 1L,
  list_fail = T
)
}
\arguments{
\item{time}{A list of time. If said time does not have a sounding, it will try to adjust to the last sounding possible. Only accepts POSIXct.}

\item{station}{Weather station code, according to the University of Wyoming website.}

\item{frost}{Include frost point calculation in dataframe. Default as \code{TRUE}.}

\item{attempt}{Attempts of downloads to try per time. Default as \code{5}.}

\item{worker}{Numbers of sessions to be open.}

\item{list_fail}{Should failed to download items be listed out? Default as \code{5}.}
}
\description{
Understanding the sounding is important for understanding the atmospheric condition, especially for extreme weather.
This catches the sounding data provided by the University of Wyoming.
}
\examples{
load_uw_asnd()
}
