% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_uw_asnd_metr.R
\name{load_uw_asnd_metr}
\alias{load_uw_asnd_metr}
\title{#' Download atmospheric sounding data}
\usage{
load_uw_asnd_metr(
  time = Sys.time(),
  station = c("45004"),
  attempt = 5L,
  worker = 1L,
  list_fail = T
)
}
\arguments{
\item{time}{A list of time. If said time does not have a sounding, it will try to adjust to the last sounding possible. Only accepts POSIXct.}

\item{station}{Weather station code, according to the University of Wyoming website.}

\item{attempt}{Attempts of downloads to try per time. Default as \code{5L}.}

\item{worker}{Numbers of sessions to be open. Default as \code{1L}.}

\item{list_fail}{Should failed to download items be listed out? Default as \code{TRUE}.}
}
\description{
Understanding the sounding is important for understanding the atmospheric condition, especially for extreme weather.
This catches index of sounding data analysised by the University of Wyoming.
}
\examples{
load_uw_asnd_metr()
}
