% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lread_csv.R
\name{lread_csv}
\alias{lread_csv}
\title{Read a list of *.csv file with the same format into a data frame}
\usage{
lread_csv(file_list)
}
\arguments{
\item{file_list}{a list of csv file directories}
}
\description{
Read a list of *.csv file with the same format into a data frame
}
\examples{
lread_csv(c("C:/Users/carol/Desktop/1.csv", "C:/Users/carol/Desktop/2.csv", "C:/Users/carol/Desktop/3.csv"))
}
