% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lread_gpx.R
\name{lread_gpx}
\alias{lread_gpx}
\title{Read a list of \code{.gpx} file with the same format into a data frame}
\usage{
lread_gpx(file, read_time = T)
}
\arguments{
\item{file}{Pathway of the files}

\item{read_time}{Should the time nodes be read and export as POSIXct in UTC? Default as \code{TRUE}.}
}
\description{
Read a list of \code{.gpx} file with the same format into a data frame
}
\examples{
lread_gpx(choose.file(), read_time = T)
}
