% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo_load_tide.R
\name{mo_load_tide}
\alias{mo_load_tide}
\title{MO weather - Download tidal images from https://www.smg.gov.mo/zh}
\usage{
mo_load_tide(
  time = weather2::tool_datetime(end = Sys.time(), duration = "91 hour", by = "1 min")
    \%>\% weather2::tool_datetime_select(by = "min", value = 0),
  dir = getwd(),
  attempt = 900,
  subattempt = 1,
  worker = 1
)
}
\arguments{
\item{time}{Date/time to download}

\item{dir}{Directory of downloaded data.}

\item{attempt}{Attempts to be made per file to download.}

\item{subattempt}{Attempts to be made per attempt to download.}
}
\value{

}
\description{
MO weather - Download tidal images from https://www.smg.gov.mo/zh
}
\examples{
mo_load_tide()
}
