% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_make.R
\name{obj_make}
\alias{obj_make}
\title{weather2 object - create a weather2 object using data}
\usage{
obj_make(data)
}
\arguments{
\item{data}{data, in a dataframe, with columns "station", "time", "type", and "value"}
}
\description{
weather2 object - create a weather2 object using data
}
\examples{
obj_make(dataframe)
}
