% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_split.R
\name{obj_split}
\alias{obj_split}
\title{weather2 object - split a weather2 object by parameter}
\usage{
obj_split(obj, by)
}
\arguments{
\item{obj}{a weather2 object}

\item{by}{parameter in question, accepts "station", "time", "type"}
}
\value{

}
\description{
weather2 object - split a weather2 object by parameter
}
\examples{
obj_split(obj, by = "station")
}
