% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_skewt.R
\name{plot_skewt}
\alias{plot_skewt}
\title{Draws the base plot of SkewT}
\usage{
plot_skewt(data, pres, skew = 45, ...)
}
\arguments{
\item{data}{Data frame itself}

\item{pres}{Column name of pressure}

\item{skew}{Skewness of the SkewT}

\item{...}{Any other parameters to be past to \code{ggplot2::aes()}}
}
\value{

}
\description{
Draws the base plot of SkewT
}
