% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_skewt_adi.R
\name{plot_skewt_adi}
\alias{plot_skewt_adi}
\title{Draws the adiabatic line of Skew-T log-P thermodynamic diagram}
\usage{
plot_skewt_adi(
  data,
  pres,
  temp,
  plim = NULL,
  dry_adi = NULL,
  wet_adi = NULL,
  isohume = NULL,
  size = 0.3,
  color = c("red", "purple", "blue")
)
}
\arguments{
\item{data}{The dataframe itself}

\item{pres}{Column name of pressure}

\item{temp}{Column name of temperature}

\item{plim}{Limit of the pressure scale. Default as \code{NULL}, which will automatically detect the maximum and minimum pressure provided in the dataframe.}

\item{dry_adi}{Values of the dry adiabatic line in Celsius, starting at the bottom of the chart.}

\item{wet_adi}{Values of the wet adiabatic line in Celsius, starting at the bottom of the chart.}

\item{isohume}{Values of the isohumes in Celsius, starting at the bottom of the chart.}

\item{size}{Size of the lines. Default as \code{0.3}.}

\item{color}{Color of the lines, ordered as dry adiabatic line, wet adiabatic line, and isohume. Default as \code{c()}}
}
\description{
Draws the adiabatic line of Skew-T log-P thermodynamic diagram
}
