% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_windstaff.R
\name{plot_windstaff}
\alias{plot_windstaff}
\title{Draws the base plot of a wind staff}
\usage{
plot_windstaff(
  data,
  pres,
  drct,
  sped,
  plim = NULL,
  pline = NULL,
  dlim = NULL,
  dline = NULL,
  ...
)
}
\arguments{
\item{data}{Data frame itself}

\item{pres}{Column name of the pressure column}

\item{drct}{Column name of the direction column}

\item{sped}{Column name of the speed column}

\item{plim}{Limit of the pressure scale. Default as \code{NULL}, which will automatically detect the maximum and minimum pressure provided in the dataframe.}

\item{pline}{Values of the Isobars. Default as \code{NULL}, which will plot the isobars at 1000, 900 ... , 200, 100, 50, 25, and 10.}

\item{dlim}{Limit of the direction scale. Default as \code{NULL}, which will automatically assume the minimum and maximum degrees of wind direction to be 0 and 360.}

\item{dline}{Values of the direction lines. Default as \code{NULL}, which will be split the \code{dlim} by 45 degrees each.}

\item{...}{}
}
\description{
Draws the base plot of a wind staff
}
