% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_csv.R
\name{read_file_csv}
\alias{read_file_csv}
\title{Read a list of *.csv file with the same format into a data frame}
\usage{
read_file_csv(file_list)
}
\arguments{
\item{file_list}{a list of csv file directories}
}
\value{

}
\description{
Read a list of *.csv file with the same format into a data frame
}
\examples{
read_file_csv(c("C:/Users/carol/Desktop/1.csv", "C:/Users/carol/Desktop/2.csv", "C:/Users/carol/Desktop/3.csv"))
}
