% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gpx.R
\name{read_gpx}
\alias{read_gpx}
\title{Read a list of \code{.gpx} file with the same format into a data frame}
\usage{
read_gpx(file, read_time = T)
}
\arguments{
\item{file}{Pathway of the files}

\item{read_time}{Should the time nodes be read and export as POSIXct in UTC? Default as \code{TRUE}.}
}
\value{

}
\description{
Read a list of \code{.gpx} file with the same format into a data frame
}
\examples{
read_gpx(choose.file(), read_time = T)
}
