% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_check_class.R
\name{sys.ck.class}
\alias{sys.ck.class}
\title{System tool: Checks the class of an object}
\usage{
sys.ck.class(value, value_name, class, all = T, silent = F)
}
\arguments{
\item{value}{Value of the object}

\item{value_name}{Name of the object, in character}

\item{class}{Expected class of an object, in character and can be a vector.}

\item{all}{Should all the values be tested in a list/ vector? Default as True}

\item{silent}{Should a text message be shown if the class of an object is incorrect? Default as False.}
}
\value{

}
\description{
System tool: Checks the class of an object
}
\examples{
sys.ck.class(1, "name", "numeric")
}
