% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_check_list.R
\name{sys.ck.list_has.na}
\alias{sys.ck.list_has.na}
\title{System tool: Check if a list includes/ excludes NA}
\usage{
sys.ck.list_has.na(list, list_name, mode = "exclude", silent = F)
}
\arguments{
\item{list}{The list}

\item{list_name}{The name of the list}

\item{mode}{Mode of comparision. Default as \code{exclude}
\itemize{
\item "exclude" (expected to have no NAs in the list)
\item "include" (expected to have NAs in the list)
}}

\item{silent}{Should a text message be shown if incorrect? Default as False.}
}
\value{

}
\description{
System tool: Check if a list includes/ excludes NA
}
\examples{
sys.ck.list_has.na(c(1, 2, 3), list_name = "obj") #Returns `TRUE`
}
