% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_check_list.R
\name{sys.ck.list_length}
\alias{sys.ck.list_length}
\title{System tool: Check if a list/vector has a certain length}
\usage{
sys.ck.list_length(list, list_name, expected, mode = "==", silent = F)
}
\arguments{
\item{list}{The list}

\item{list_name}{The name of the list}

\item{expected}{Expected length of the list. Must be an integer}

\item{mode}{Mode of comparison. Default as \code{==}
\itemize{
\item "==" (equal to)
\item "!=" (not equal to)
\item ">" (larger than)
\item "<" (smaller than)
\item ">=" (larger or equal to)
\item "<=" (smaller or equal to)
}}

\item{silent}{Should a text message be shown if incorrect? Default as False.}
}
\value{

}
\description{
System tool: Check if a list/vector has a certain length
}
\examples{
sys.ck.list_length(c(1,2,3), "obj", 2L) #Return `TRUE`
}
