% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_check_list.R
\name{sys.ck.list_numericable}
\alias{sys.ck.list_numericable}
\title{System tool: Check if a list can be turned into a numeric}
\usage{
sys.ck.list_numericable(list, list_name, silent = F)
}
\arguments{
\item{list}{The list}

\item{list_name}{The name of the list}

\item{silent}{Should a text message be shown if incorrect? Default as False.}
}
\value{

}
\description{
System tool: Check if a list can be turned into a numeric
}
\examples{
sys.ck.list_numericable(list = c(1, 1, "a"), list_name = "obj") #Returns `TRUE`
}
