% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_load_file.R
\name{sys.load_file.seq}
\alias{sys.load_file.seq}
\title{System tools: Download files from website}
\usage{
sys.load_file.seq(
  data,
  title = "test_seq",
  attempt = 5,
  threshold = 0.5,
  list_fail = T
)
}
\arguments{
\item{data}{Data frame containing columns "URL", "DIR", "Info"}

\item{title}{Title of the downloaded data}

\item{attempt}{Attempts to be made per download.}

\item{threshold}{Threshold of file size, when compared to the largest item to be installed in the list}

\item{list_fail}{List failed-to-download items}
}
\description{
Warning: This is not a "checked" function. Your input, if incorrect, may damage your system.
Please use the sys.load_file() function instead.
}
\examples{
sys.load_file.seq(data, title = "Global nc data")
}
\keyword{internal}
