% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_load_fileset.R
\name{sys.load_fileset}
\alias{sys.load_fileset}
\title{System tools: Download files from website}
\usage{
sys.load_fileset(
  data,
  title = "test",
  attempt = 5,
  worker = 0,
  list_fail = T,
  threshold = 0.5,
  check = T
)
}
\arguments{
\item{data}{Data frame containing columns "URL", "DIR", "Info", "Set"}

\item{title}{Title of the downloaded data}

\item{attempt}{Attempts to be made per download URL to be attempted}

\item{worker}{Numbers of sessions to be open}

\item{list_fail}{List failed-to-download items}

\item{threshold}{Threshold of file size, when compared to the largest item to be installed in the list}

\item{check}{Check the all of the parameters before execution, default as T}
}
\value{

}
\description{
System tools: Download files from website
}
\examples{
sys.load_fileset(data, title = "Macao tidal data", attempt = 10, worker = 20, list_fail = T, threshold = 0.5)
}
