% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_tool_data.R
\name{sys.tl.data_get.colname}
\alias{sys.tl.data_get.colname}
\title{System tools: Get a non-clashing column name within a dataframe, by adding a numeric suffix ("_x")}
\usage{
sys.tl.data_get.colname(value, data)
}
\arguments{
\item{value}{Column name you are expecting to use. Can be a symbol or a character.}

\item{data}{The dataframe you are using}
}
\value{
A character
}
\description{
System tools: Get a non-clashing column name within a dataframe, by adding a numeric suffix ("_x")
}
\examples{
sys.tl.data_get.colname(x, data = tibble::tibble(x = 0)) #Returns "x_1"
}
