% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_check_class.R
\name{sys_ckc_list}
\alias{sys_ckc_list}
\title{System tool: Checks an object is a list}
\usage{
sys_ckc_list(value, value_name, all = F, silent = F)
}
\arguments{
\item{value}{Value of the object}

\item{value_name}{Name of the object, in character}

\item{all}{Should all the values be tested in a list/ vector? Default as True}

\item{silent}{Should a text message be shown if the class of an object is incorrect? Default as False.}
}
\value{

}
\description{
A wrapper around \code{sys_ckc}. \strong{Note that by default, \code{all} is set as \code{FALSE}}
}
\examples{
sys_ckc_list(list(c(1,2,3), c(3, 2, 1)), "obj")
}
