% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_check_class.R
\name{sys_ckc_numeric}
\alias{sys_ckc_numeric}
\title{System tool: Checks an object is a numeric}
\usage{
sys_ckc_numeric(value, value_name, all = T, silent = F)
}
\arguments{
\item{value}{Value of the object}

\item{value_name}{Name of the object, in character}

\item{all}{Should all the values be tested in a list/ vector? Default as True}

\item{silent}{Should a text message be shown if the class of an object is incorrect? Default as False.}

\item{class}{Expected class of an object, in character and can be a vector.}
}
\description{
A wrapper around \code{sys_ckc}
}
\examples{
sys_ckc_numeric(1, "obj")
}
