% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_check_data.R
\name{sys_ckd_colexist}
\alias{sys_ckd_colexist}
\title{System tool: Check if a colname is present/ absent in a dataframe}
\usage{
sys_ckd_colexist(
  value,
  value_name,
  data,
  data_name,
  mode = "present",
  silent = F
)
}
\arguments{
\item{value}{The value. Accpets \code{symbols} and \code{character}. Only accepts 1.}

\item{value_name}{The name of value}

\item{data}{The dataframe}

\item{data_name}{The name of dataframe}

\item{mode}{Mode of comparison. Default as \code{present}
\itemize{
\item "present" (\code{value} should be a column within \code{data})
\item "absent" (\code{value} should \strong{not} be a column within \code{data})
}}

\item{silent}{Should a text message be shown if incorrect? Default as \code{False}.}
}
\value{

}
\description{
System tool: Check if a colname is present/ absent in a dataframe
}
\examples{
sys_ckd_colexist(value = y, value_name = "x", data = tibble::tibble(y = 0), data_name = "hi", mode = "absent") #Returns `TRUE`
}
