% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_check_data.R
\name{sys_ckd_grouped}
\alias{sys_ckd_grouped}
\title{System tool: Check if grouping is present/ absent in a dataframe}
\usage{
sys_ckd_grouped(data, data_name, mode = "absent", silent = F)
}
\arguments{
\item{data}{The dataframe}

\item{data_name}{The name of dataframe}

\item{mode}{Mode of comparison. Default as \code{present}
\itemize{
\item "present" (\code{data} should be grouped)
\item "absent" (\code{data} should \strong{not} be grouped)
}}

\item{silent}{Should a text message be shown if incorrect? Default as \code{False}.}
}
\value{
A logical value. \code{TRUE} if an error occured. \code{FALSE} if no error occured. A text message is provided if \code{silent} is \code{FALSE}.
}
\description{
System tool: Check if grouping is present/ absent in a dataframe
}
\examples{
sys_ckd_grouped(data(iris))
}
