% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_check_fn.R
\name{sys_ckf_CalcSmooth}
\alias{sys_ckf_CalcSmooth}
\title{System tool: Check if the input satisfy the \code{calc_smooth_} requirements}
\usage{
sys_ckf_CalcSmooth(data, based, value)
}
\arguments{
\item{data}{The dataframe to be smoothened}

\item{based}{The column name of the explanatory variable}

\item{value}{The column name of the response variable}
}
\value{

}
\description{
System tool: Check if the input satisfy the \code{calc_smooth_} requirements
}
