% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_check_fn.R
\name{sys_ckf_NameAsReturn}
\alias{sys_ckf_NameAsReturn}
\title{System tool: Check if the input satisfy the \code{sys_tld_FormatReturn} requirements}
\usage{
sys_ckf_NameAsReturn(name_as, overwrite, expected)
}
\arguments{
\item{name_as}{Names of the newly added columns into the dataframe. The keyword "\emph{del}" is used here.}

\item{overwrite}{Let the new column names to overwrite the original dataframe columns?}

\item{expected}{Expected length of \code{name_as}, in integer.}
}
\value{
A logical value, with \code{TRUE} as encountering an issue with input, and \code{FALSE} as successful. Prints error message when \code{TRUE}.
}
\description{
System tool: Check if the input satisfy the \code{sys_tld_FormatReturn} requirements
}
\examples{
sys_ckf_NameAsReturn("temp2", TRUE, 1L)
}
