% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_check_fn.R
\name{sys_ckf_SysLoadFile}
\alias{sys_ckf_SysLoadFile}
\title{System tool: Check if the input satisfy the sys_load_file requirements}
\usage{
sys_ckf_SysLoadFile(data, title, attempt, worker, list_fail, threshold)
}
\arguments{
\item{data}{Data frame containing columns "URL", "DIR", "Info"}

\item{title}{Title of the downloaded data}

\item{attempt}{Attempts to be made per download.}

\item{worker}{Numbers of sessions to be open}

\item{list_fail}{List failed-to-download items}

\item{threshold}{Threshold of file size, when compared to the largest item to be installed in the list}
}
\description{
System tool: Check if the input satisfy the sys_load_file requirements
}
