% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_check_fn.R
\name{sys_ckf_SysLoadLine}
\alias{sys_ckf_SysLoadLine}
\title{System tool: Check if the input satisfy the \code{sys_load_line} requirements}
\usage{
sys_ckf_SysLoadLine(data, title, attempt, worker, list_fail, threshold)
}
\arguments{
\item{data}{Dataframe containing columns \code{URL} and \code{Info}.}

\item{title}{Title of the downloaded data}

\item{attempt}{Attempts to be made per download URL to be attempted.}

\item{worker}{Numbers of sessions to be open.}

\item{list_fail}{List failed-to-download items}

\item{threshold}{}
}
\description{
System tool: Check if the input satisfy the \code{sys_load_line} requirements
}
\examples{
sys_ckf_SysLoadLine(data, "title", 5L, 20L, TRUE, 1L)
}
