% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_check_list.R
\name{sys_ckl_ItemIn}
\alias{sys_ckl_ItemIn}
\title{System tool: Check if items in a list are in/ not in some expected value}
\usage{
sys_ckl_ItemIn(list, list_name, expected, mode = "in", silent = F)
}
\arguments{
\item{list}{The list}

\item{list_name}{The name of the list}

\item{expected}{Expected values in the list}

\item{mode}{Mode of comparision. Default as \verb{in}
\itemize{
\item "in" (within a list of expected value)
\item "out" (\strong{not} within a list )
}}

\item{silent}{Should a text message be shown if incorrect? Default as False.}
}
\description{
System tool: Check if items in a list are in/ not in some expected value
}
\examples{
sys_ckl_ItemIn(c(1, 2, 3, 4), "obj", c("1", "2", "3"), mode = "out") #Returns `TRUE`
}
