% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_check_list.R
\name{sys_ckl_ItemUnique}
\alias{sys_ckl_ItemUnique}
\title{System tool: Check if items in list are unique or common}
\usage{
sys_ckl_ItemUnique(list, list_name, mode = "all unique", silent = F)
}
\arguments{
\item{list}{The list}

\item{list_name}{The name of the list}

\item{mode}{Mode of comparison. Default as \code{unique}
\itemize{
\item "all unique" (expecting all items are unique items in the list)
\item "common" (expected some items are not unique items in the list)
}}

\item{silent}{Should a text message be shown if incorrect? Default as False.}
}
\value{

}
\description{
System tool: Check if items in list are unique or common
}
\examples{
sys_ckl_ItemUnique(list = c(1,2,3), "obj", mode = "common") #Returns `TRUE`
}
