% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_check_list.R
\name{sys_ckl_NumericValue}
\alias{sys_ckl_NumericValue}
\title{System tool: Check if items in a numeric/integer list are of an expected value}
\usage{
sys_ckl_NumericValue(list, list_name, expected, mode = "==", silent = F)
}
\arguments{
\item{list}{The list}

\item{list_name}{Name of \code{list}}

\item{expected}{Expected value in the list}

\item{mode}{Mode of comparison. Default as \code{==}
\itemize{
\item "==" (values equal to the expected value)
\item "!=" (values not equal to expected value)
\item ">" (values larger than expected value)
\item "<" (values smaller than expected value)
\item ">=" (values larger or equal to expected value)
\item "<=" (values smaller or equal to expected value)
}}

\item{silent}{Should a text message be shown if incorrect? Default as \code{False}.}
}
\value{

}
\description{
System tool: Check if items in a numeric/integer list are of an expected value
}
