% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_load_line.R
\name{sys_load_lineSEQ}
\alias{sys_load_lineSEQ}
\title{System tools: Download lines from website}
\usage{
sys_load_lineSEQ(
  data,
  title = "test_seq",
  attempt = 5,
  threshold = 1,
  list_fail = T
)
}
\arguments{
\item{data}{Data frame containing columns \code{URL} and \code{Info}.}

\item{title}{Title of the downloaded data}

\item{attempt}{Attempts to be made per download URL to be attempted. Default as a maximum of \code{5} times.}

\item{threshold}{Threshold of line size. Any downloaded lines with number less than the threshold will be considered as failure and will be reattempted. Default as a minimum of \code{1} line.}

\item{list_fail}{List failed-to-download items. Default as \code{TRUE}.}
}
\description{
\strong{Warning: This is not a "checked" function. Your input, if incorrect, may damage your system.}
\strong{Please use the sys_load_line() function instead.}
}
\examples{
sys_load_lineSEQ(data, title = "Wyoming Atmospheric Sounding Data")
}
\keyword{internal}
