% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_tool_data.R
\name{sys_tld_FormatReturn}
\alias{sys_tld_FormatReturn}
\title{System tools: Return data by \code{name_as} and \code{overwrite}, with keyword \code{"*del*"}}
\usage{
sys_tld_FormatReturn(data, name_as, value, overwrite = F)
}
\arguments{
\item{data}{The dataframe}

\item{name_as}{Names of the newly added columns into the dataframe. The keyword \code{"*del*"} is used here.}

\item{value}{Values of the new columns in a list.
\strong{NOTE}
\itemize{
\item The list must have the same length as \code{name_as}
\item The items in the list must have the same length as the rows numbers of \code{data}
}}

\item{overwrite}{Let the new column names to overwrite the original dataframe columns? Default as \code{FALSE}.}
}
\description{
The special keyword \code{"*del*"} allows user to not add specific columns into the returned dataframe, even if the column is calculated.
}
