% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_tool_data.R
\name{sys_tld_GrpByAltVal}
\alias{sys_tld_GrpByAltVal}
\title{System tools: Return groups of data, grouped by shifting values and row numbers}
\usage{
sys_tld_GrpByAltVal(data, col, name_as = "grp", overwrite = F)
}
\arguments{
\item{data}{The dataframe itself}

\item{col}{The selected column of values in the dataframe}

\item{name_as}{Name of the new smoothed column. Default as \code{"grp"}. Accepts the keyword \code{"*del*"}}

\item{overwrite}{Let the new column names to overwrite the original dataframe columns? Default as \code{FALSE}.}
}
\description{
In ascending order of row number, a sudden shift in value of the selected column will add the group number by 1.
The group number will always start at 1.
}
\examples{
tibble::tibble(val = c(1,1,1,1,-1,-1, 1,1)) \%>\% sys_tld_GrpByAltVal(val)
}
