% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_asnd_dict.R
\name{tool_asnd_dict}
\alias{tool_asnd_dict}
\title{Tool: Provide dictionary of commonly used sounding index abbreviation/ codes}
\usage{
tool_asnd_dict()
}
\description{
Tool: Provide dictionary of commonly used sounding index abbreviation/ codes
}
\examples{
tool_asnd_dict()
}
