% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_date.R
\name{tool_date}
\alias{tool_date}
\title{Tool: To create a sequence of date}
\usage{
tool_date(
  start,
  end = as.Date(lubridate::with_tz(Sys.time(), tzone = "UTC")),
  by = "1 days",
  duration = "1 days"
)
}
\arguments{
\item{start}{Starting time, can be in the format of character or date}

\item{end}{Ending time, can be in the format of character or date}

\item{by}{Time interval of the sequence. Must be a string.}

\item{duration}{Duration between starting time and ending time. Must be a string. Overrides when both start and end is supplied.}
}
\description{
Tool: To create a sequence of date
}
\examples{
tool_datetime(end = Sys.time(), by = "1 day", duration = "30 day")
}
