% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_datetime_select.R
\name{tool_datetime_select}
\alias{tool_datetime_select}
\title{Tool: To select date time by a value}
\usage{
tool_datetime_select(time, by, value)
}
\arguments{
\item{time}{A sequence of POSIXct}

\item{by}{A character string of "sec", "min", "hour", "day", "month", or "year"}

\item{value}{Value of time selected by "by". Can be a vector.}
}
\value{

}
\description{
Tool: To select date time by a value
}
\examples{
tool_datetime_select(time, "min", 0)
}
