% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_hk_wxph_dict.R
\name{tool_hk_wxph}
\alias{tool_hk_wxph}
\title{Tool: Provide dictionary of HKO' weather camera station code}
\usage{
tool_hk_wxph()
}
\value{

}
\description{
Tool: Provide dictionary of HKO' weather camera station code
}
\examples{
tool_hk_wxph()
}
