% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_metar.R
\name{tool_metar}
\alias{tool_metar}
\title{Tool: Decode METAR reports}
\usage{
tool_metar(data, code, CAVOK)
}
\arguments{
\item{data}{Dataframe that contains METAR code}

\item{code}{The column name of the METAR code}

\item{CAVOK}{If decode is T, supress CAVOK and convert all CAVOK into vsby = 9999. Default as T.}
}
\value{

}
\description{
Tool: Decode METAR reports
}
\examples{
tool_metar(data, code)
}
