% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_metar_dict.R
\name{tool_metar_dict}
\alias{tool_metar_dict}
\title{Tool: Provide dictionary of commonly used METAR abbreviation/ codes}
\usage{
tool_metar_dict(type = "weather")
}
\arguments{
\item{type}{Type of dictionary to provide. Only accepts "weather" or "cloud", and only 1 input.}
}
\value{

}
\description{
Tool: Provide dictionary of commonly used METAR abbreviation/ codes
}
\examples{
tool_metar_dict("weather")
}
