% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_check_col_exist.R
\name{w2_check_col_exist}
\alias{w2_check_col_exist}
\title{System tool: Check if an object is a column name under data}
\usage{
w2_check_col_exist(data, value, value_name)
}
\arguments{
\item{data}{Data}

\item{value}{Value of the check item}

\item{value_name}{Value's name as a string}
}
\value{

}
\description{
System tool: Check if an object is a column name under data
}
\examples{
w2_check_col_exist(data, info, "info")
}
