% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_check_int.R
\name{w2_check_int}
\alias{w2_check_int}
\title{System tool: Check if a value is a integer}
\usage{
w2_check_int(value, value_name)
}
\arguments{
\item{value}{value}

\item{value_name}{name of the value}
}
\value{

}
\description{
System tool: Check if a value is a integer
}
\examples{
w2_check_int(attempt, "attempt")
}
