% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_check_internet.R
\name{w2_check_internet}
\alias{w2_check_internet}
\title{System tool: Check if internet is connected}
\usage{
w2_check_internet(
  attempt = 5,
  silent = T,
  websites = c("https://google.com", "https://www.un.org/")
)
}
\arguments{
\item{attempt}{Attempts of internet check to be made. Default as 5.}

\item{silent}{Silent successful message}

\item{websites}{Website to test}
}
\value{

}
\description{
System tool: Check if internet is connected
}
\examples{
w2_check_internet()
}
