% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_check_item_value.R
\name{w2_check_item_value}
\alias{w2_check_item_value}
\title{System tool: Check if a numeric value is equal to expected value}
\usage{
w2_check_item_value(item, item_name, expect, type = "=")
}
\arguments{
\item{item}{item}

\item{item_name}{name of item}

\item{expect}{expected value}

\item{type}{type}
}
\value{

}
\description{
System tool: Check if a numeric value is equal to expected value
}
