% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_check_lan.R
\name{w2_check_lan}
\alias{w2_check_lan}
\title{System tool: Check if a value is a language flag}
\usage{
w2_check_lan(value, value_name)
}
\arguments{
\item{value}{value}

\item{value_name}{name of the value}
}
\value{

}
\description{
System tool: Check if a value is a language flag
}
\examples{
w2_check_lan(lan, "lan")
}
