% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_check_length.R
\name{w2_check_length}
\alias{w2_check_length}
\title{System tool: Check if a value has a certain length}
\usage{
w2_check_length(value, length, value_name)
}
\arguments{
\item{value}{value}

\item{length}{expected length of the value}

\item{value_name}{name of the value}
}
\value{

}
\description{
System tool: Check if a value has a certain length
}
\examples{
w2_check_length(c("sec", "min"), 1, "by")
}
