% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_check_list_item.R
\name{w2_check_list_item}
\alias{w2_check_list_item}
\title{System tool: Check if an item is within a list}
\usage{
w2_check_list_item(item, item_name, list)
}
\arguments{
\item{item}{item}

\item{item_name}{name of the item}

\item{list}{list}
}
\value{

}
\description{
System tool: Check if an item is within a list
}
