% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_check_list_na.R
\name{w2_check_list_na}
\alias{w2_check_list_na}
\title{System tool: Check if a list contains NA}
\usage{
w2_check_list_na(data, data_name, NAs = F)
}
\arguments{
\item{data}{data}

\item{data_name}{name of the data}

\item{NAs}{check if it should contains NAs(T) or not contain NAs(F)}
}
\value{

}
\description{
System tool: Check if a list contains NA
}
