% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_check_list_numericable.R
\name{w2_check_list_numericable}
\alias{w2_check_list_numericable}
\title{System tool: Check if a list can be turned into a numeric}
\usage{
w2_check_list_numericable(data, data_name)
}
\arguments{
\item{data}{data}

\item{data_name}{name of the data}
}
\value{

}
\description{
System tool: Check if a list can be turned into a numeric
}
