% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_check_list_unique.R
\name{w2_check_list_unique}
\alias{w2_check_list_unique}
\title{System tool: Check if a list are all unique items}
\usage{
w2_check_list_unique(data, data_name)
}
\arguments{
\item{data}{data}

\item{data_name}{name of the data}
}
\value{

}
\description{
System tool: Check if a list are all unique items
}
