% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_check_posixct.R
\name{w2_check_posixct}
\alias{w2_check_posixct}
\title{System tool: Check if a value is a POSIXct}
\usage{
w2_check_posixct(value, value_name)
}
\arguments{
\item{value}{value}

\item{value_name}{name of the value}
}
\value{

}
\description{
System tool: Check if a value is a POSIXct
}
\examples{
w2_check_posixct(time, "time")
}
