% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_check_str.R
\name{w2_check_str}
\alias{w2_check_str}
\title{System tool: Check if a value is a string}
\usage{
w2_check_str(value, value_name)
}
\arguments{
\item{value}{value}

\item{value_name}{name of the value}
}
\value{

}
\description{
System tool: Check if a value is a string
}
\examples{
w2_check_str(title, "title")
}
