% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_check_type_dataframe.R
\name{w2_check_type_dataframe}
\alias{w2_check_type_dataframe}
\title{System tool: Check if an object is a dataframe}
\usage{
w2_check_type_dataframe(data, data_name)
}
\arguments{
\item{data}{data}

\item{data_name}{name of the data}
}
\value{

}
\description{
System tool: Check if an object is a dataframe
}
