% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_check_type.R
\name{w2_check_type_help}
\alias{w2_check_type_help}
\title{System tool: Checks the input of all w2_check_type_xxx function}
\usage{
w2_check_type_help(value, value_name)
}
\arguments{
\item{value_name}{}
}
\value{

}
\description{
System tool: Checks the input of all w2_check_type_xxx function
}
