% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_check_type_integer.R
\name{w2_check_type_integer}
\alias{w2_check_type_integer}
\title{System tool: Check if an object is an integer}
\usage{
w2_check_type_integer(value, value_name)
}
\arguments{
\item{value}{value}

\item{value_name}{name of the value}
}
\value{

}
\description{
System tool: Check if an object is an integer
}
