% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_check_type.R
\name{w2_check_type_POSIXct}
\alias{w2_check_type_POSIXct}
\title{System tool: Check if an object is a POSIXct}
\usage{
w2_check_type_POSIXct(value, value_name)
}
\arguments{
\item{value}{value}

\item{value_name}{name of the value}
}
\value{

}
\description{
System tool: Check if an object is a POSIXct
}
