% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_col_exist.R
\name{w2_col_exist}
\alias{w2_col_exist}
\title{System tool: Check if a column exist in data}
\usage{
w2_col_exist(data, col)
}
\arguments{
\item{data}{Data}

\item{col}{Column name. Doesn't need to be in "".}
}
\value{

}
\description{
System tool: Check if a column exist in data
}
\examples{
w2_col_exist(data, info)
}
