% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_get_colname.R
\name{w2_get_colname}
\alias{w2_get_colname}
\title{System tools: Get a non-clashing column name within a dataframe, by adding a numeric suffix ("_x")}
\usage{
w2_get_colname(data, name)
}
\arguments{
\item{data}{Dataframe}

\item{name}{Column name you are expecting to use}
}
\value{

}
\description{
System tools: Get a non-clashing column name within a dataframe, by adding a numeric suffix ("_x")
}
\examples{
w2_get_colname(data, "predict")
}
