% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_load_file_seq.R
\name{w2_load_file_seq}
\alias{w2_load_file_seq}
\title{System tools: Download files from website}
\usage{
w2_load_file_seq(data, title, list_fail = F, attempt = 10)
}
\arguments{
\item{data}{Data frame containing columns "URL", "DIR", "Info"}

\item{title}{Title of the downloaded data}

\item{list_fail}{List failed-to-download items}

\item{attempt}{Attempts to be made per download.}
}
\value{

}
\description{
System tools: Download files from website
}
\examples{
w2_load_file_seq(data, title = "Global nc data", attempt = 10)
}
