% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_load_help_formatdata.R
\name{w2_load_help_formatdata}
\alias{w2_load_help_formatdata}
\title{System tools: Format data with exist and size column in any w2_load_... functions}
\usage{
w2_load_help_formatdata(data)
}
\arguments{
\item{data}{Data frame containing columns "URL", "DIR", "Info"}
}
\value{

}
\description{
System tools: Format data with exist and size column in any w2_load_... functions
}
\examples{
w2_load_help_formatdata(data)
}
