% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_load_help_listfail.R
\name{w2_load_help_listfail}
\alias{w2_load_help_listfail}
\title{System tools: Message failed items in any w2_load_... functions}
\usage{
w2_load_help_listfail(list_of_fail)
}
\arguments{
\item{list_of_fail}{List of the failed items, in a character vector}
}
\value{

}
\description{
System tools: Message failed items in any w2_load_... functions
}
\examples{
w2_load_help_listfail(fail$Info)
}
