% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_td_ema.R
\name{calc_ema}
\alias{calc_ema}
\title{Calculate Emagram plot coordinates}
\usage{
calc_ema(data, x, y, name_as = NULL, overwrite = F)
}
\arguments{
\item{angle}{}
}
\description{
Calculate Emagram plot coordinates
}
