% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_td_ema.R
\name{calc_emax}
\alias{calc_emax}
\title{Calculate Emagram plot x-coordinate}
\usage{
calc_emax(data, x, name_as = NULL)
}
\arguments{
\item{name_as}{}
}
\description{
Calculate Emagram plot x-coordinate
}
\examples{


}
\keyword{internal}
