% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_td_ema.R
\name{calc_emay}
\alias{calc_emay}
\title{Calculate Emagram plot y-coordinate}
\usage{
calc_emay(data, y, name_as = NULL)
}
\arguments{
\item{data}{The dataframe iteself.}

\item{y}{The column name of y-coordinate to be plotted onto the emagram. Usually it's some form of "pressure" data.}

\item{name_as}{Names of the 1 new columns, i.e. the y-coordinate of the emagram. Default as \code{NULL}, i.e. the column name of \code{x} with a suffix of \code{"_emx"}. Keyword \code{"*del*"} is supported.}
}
\value{
The same dataframe as \code{data}, with 1 additional column with column name defined in \code{name_as}.
}
\description{
Emagrams are one of the 5 meteorological thermo-diagram, which in essence is a log-P linear-T diagram.
This function calculates the emagram \strong{Y} coordinates in a linear graph.
\strong{Warning: Internal function. Inputs are not checked before executed. Use with care.}
\strong{Warning: This function has over_write function set as \code{TRUE}}
}
\examples{
calc_emax(data, pres)
}
\keyword{internal}
