% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lcl_est.R
\name{calc_lcl_est}
\alias{calc_lcl_est}
\title{Calculate the \strong{estimated} final temperature and pressure of Lifting Condensation Level}
\usage{
calc_lcl_est(
  data,
  pres,
  temp,
  dwpt,
  name_as = c("lclp", "lclt_t", "lclt_d"),
  overwrite = F
)
}
\arguments{
\item{data}{The dataframe itself. Each row of data will be considered as an observation for each calculation}

\item{pres}{The column name of the initial pressure of the calculation. In any units.}

\item{temp}{The column name of the initial temperature of the calculation. Unit in \code{degC}.}

\item{dwpt}{The column name of the initial dewpoint of the calculation. Unit in \code{degC}.}

\item{name_as}{Names of the 3 new columns, in the order of "pressure of the LCL", "temperature of the LCL(estimated by temperature)", and "temperature of the LCL(estimated by dewpoint)". Default as \code{c("lclp", "lclt_t", "lclt_d")}.}

\item{overwrite}{Let the new column names to overwrite the original dataframe columns? Default as \code{FALSE}.}
}
\value{

}
\description{
Calculate the \strong{estimated} final temperature and pressure of Lifting Condensation Level
}
\examples{
calc_lcl_est(tibble::tibble(pres = 1000, temp = 20, dwpt = 10), pres, temp, dwpt) #Should return 861 for pressure
}
