% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_td_skewt.R
\name{calc_skewt}
\alias{calc_skewt}
\title{Calculate Skew-T plot coordinates}
\usage{
calc_skewt(data, x, y, mode = NULL, angle = 45)
}
\arguments{
\item{angle}{}
}
\value{

}
\description{
Calculate Skew-T plot coordinates
}
