% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_td_skewt.R
\name{calc_skewtx}
\alias{calc_skewtx}
\title{Calculate Skew-T plot x-coordinate}
\usage{
calc_skewtx(data, x, y, angle, name_as = NULL)
}
\arguments{
\item{name_as}{}
}
\description{
Calculate Skew-T plot x-coordinate
}
\examples{


}
\keyword{internal}
