% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_td_skewt.R
\name{calc_skewty}
\alias{calc_skewty}
\title{Calculate Skew-T plot y-coordinate}
\usage{
calc_skewty(data, y, name_as = NULL)
}
\arguments{
\item{name_as}{}
}
\description{
Calculate Skew-T plot y-coordinate
}
\keyword{internal}
