% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_smooth_ma.R
\name{calc_smooth_ma}
\alias{calc_smooth_ma}
\title{Calculate by smoothing the column, by moving average}
\usage{
calc_smooth_ma(
  data,
  based,
  value,
  type = "center",
  weight = 3,
  name_as = "",
  NAs = T
)
}
\arguments{
\item{data}{The Data frame}

\item{based}{A column within dataframe. This column must be "numeric-able", must not contain NAs, and must be unique}

\item{value}{A column within data}

\item{type}{Accepts 1 string of "left", "center", or "right"}

\item{weight}{Weighting or the size of the window}

\item{name_as}{Name of the new smoothed column. Default as predict_xxx, where xxx is the column name of the based.}

\item{NAs}{Removes NA in the smoothed column by treating missing data}
}
\value{

}
\description{
Calculate by smoothing the column, by moving average
}
