% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_smooth_sp.R
\name{calc_smooth_sp}
\alias{calc_smooth_sp}
\title{Calculate by smoothing the column, by smoothing spline}
\usage{
calc_smooth_sp(data, based, value, df, name_as = "")
}
\arguments{
\item{data}{The Data frame}

\item{based}{A column within dataframe. This column must be "numeric-able", must not contain NAs, and must be unique}

\item{value}{A column within data}

\item{df}{Degree of freedom. Must be between 1 and the number of rows in data (rows with NA values excluded)}

\item{name_as}{Name of the new smoothed column. Default as predict_xxx, where xxx is the column name of the based.}
}
\value{

}
\description{
Calculate by smoothing the column, by smoothing spline
}
\examples{
calc_smooth_sp(data, x, y, df = 0.7*nrow(data), name_as = "")
}
