% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_time_int.R
\name{calc_time_int}
\alias{calc_time_int}
\title{Calculate time interval, between previous or next row}
\usage{
calc_time_int(data, time, type = "before", unit = "minute", name_as)
}
\arguments{
\item{data}{Dataframe or Weather containing POSIXct time}

\item{time}{Column name of the time}

\item{type}{Type of operation, accepts "before" or "after" only.}

\item{unit}{Unit of time interval. Default as "minute".}

\item{name_as}{Column name of the time interval. If not provided, default as time with suffix "_int".}
}
\value{

}
\description{
Calculate time interval, between previous or next row
}
\examples{
calc_time_int(data, time)
}
