% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_td_skewt.R
\name{draw_td_skewt}
\alias{draw_td_skewt}
\title{Draws a blank Skew-T plot}
\usage{
draw_td_skewt(
  data,
  pres = c(seq(1000, 100, -100), 50, 25, 10),
  temp = seq(-120, 50, 10),
  dry_adi = seq(-100, 100, 10),
  wet_adi = seq(-100, 100, 10),
  isohume = seq(-100, 100, 10),
  angle = 45,
  size = 10,
  pres_lim = NULL,
  temp_lim = c(-80, 50)
)
}
\arguments{
\item{temp_lim}{}
}
\value{

}
\description{
Draws a blank Skew-T plot
}
