% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_info.R
\name{file_info}
\alias{file_info}
\title{Extract file information with path}
\usage{
file_info(file)
}
\arguments{
\item{file}{list of file paths}
}
\description{
Basically a wrapper for base::file.info
}
\examples{
file_info(c("1.csv", "2.csv", "3.csv"))
}
