% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_skewt.R
\name{plot_skewt}
\alias{plot_skewt}
\title{Draws the base plot of Skew-T log-P thermodynamic diagram}
\usage{
plot_skewt(
  data,
  pres,
  temp,
  skew = 45,
  plim = NULL,
  pline = NULL,
  tlim = NULL,
  tline = NULL,
  ...
)
}
\arguments{
\item{data}{Data frame itself}

\item{pres}{Column name of pressure}

\item{skew}{Skewness of the Skew-T}

\item{plim}{Limit of the pressure scale. Default as \code{NULL}, which will automatically detect the maximum and minimum pressure provided in the dataframe.}

\item{pline}{Values of the Isobars. Default as \code{NULL}, which will plot the isobars at 1000, 900 ... , 200, 100, 50, 25, and 10.}

\item{tline}{Values of the Isotherms. Default as \code{NULL}, which will plot the isotherms automatically the range of the provided temperature profile, with some to spare.}

\item{...}{Any other parameters to be past to \code{ggplot2::aes()}}
}
\description{
Draws the base plot of Skew-T log-P thermodynamic diagram
}
