% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_check_col_exist.R
\name{w2_check_col_exist}
\alias{w2_check_col_exist}
\title{System tool: Check if a column exists in a dataframe}
\usage{
w2_check_col_exist(data, data_name, value, value_name)
}
\arguments{
\item{data}{data}

\item{data_name}{name of the data}

\item{value}{value}

\item{value_name}{name of the value}
}
\value{

}
\description{
System tool: Check if a column exists in a dataframe
}
