% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_load_fileset_seq.R
\name{w2_load_fileset_seq}
\alias{w2_load_fileset_seq}
\title{System tools: Download files from website}
\usage{
w2_load_fileset_seq(data, title, subattempt = 1)
}
\arguments{
\item{data}{Data frame containing columns "URL", "DIR", "Info"}

\item{title}{Title of the downloaded data}

\item{subattempt}{Attempts to be made per download URL to be attempted.}
}
\value{

}
\description{
System tools: Download files from website
}
\examples{
w2_load_fileset_seq(data, title = "Macao tidal data", subattempt = 10)
}
