% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_load_fileset_stm.R
\name{w2_load_fileset_stm}
\alias{w2_load_fileset_stm}
\title{System tools: Download files from website using STORM mode}
\usage{
w2_load_fileset_stm(data, title, subattempt = 1, worker = 20)
}
\arguments{
\item{data}{Data frame containing columns "URL", "DIR", "Info"}

\item{title}{Title of the downloaded data}

\item{subattempt}{Attempts to be made per download URL to be attempted.}

\item{worker}{Numbers of sessions to be open}
}
\value{

}
\description{
System tools: Download files from website using STORM mode
}
\examples{
w2_load_fileset_stm(data, title = "Macao tidal data", subattempt = 10, worker = 10)
}
